postprocess linebreak and non-linebreak MathML: https://github.com/metanorma/metanorma-standoc/issues/111
